reorganized js so html files now will only need the init js file to run
