+++ error.html
@@ -76,8 +76,6 @@
       <p id="copyright"></p>
     </div>
   </footer>
-  <script type="module" src="/js/components/_form.js"></script>
   <script type="module" src="/js/init/error.js"></script>
-
 </body> 
 </html>
