+++ js/render/sign-in-page/sign-in.js
@@ -1,3 +1,4 @@
+import { formComponent } from "../../components/_form.js";
 import { menuNavComponent } from "../../components/_menu-nav.js";
 import { renderCopyright } from "../shared/copyright.js";
 import { renderSignInForm } from "./sign-in-form.js";
@@ -5,6 +6,7 @@ import { renderSignInForm } from "./sign-in-form.js";
 function initSignInPage() {
     menuNavComponent();
     renderSignInForm();
+    formComponent();
     renderCopyright();
 }
