+++ movie.html
@@ -94,8 +94,6 @@
     </div>
   </footer>
   <script src="https://cdn.jsdelivr.net/npm/swiper@9/swiper-bundle.min.js"></script>
-  <script type="module" src="/js/components/_form.js"></script>
-  <script type="module" src="/js/render/shared/modals.js"></script>
   <script type="module" src="/js/init/movie.js"></script>
 </body>
 </html>
