+++ sign-in.html
@@ -59,7 +59,6 @@
       <p id="copyright"></p>
     </div>
   </footer>
-  <script type="module" src="/js/components/_form.js"></script>
   <script type="module" src="/js/init/sign-in.js"></script>
 </body> 
 </html>
